Implement simplest ls with opendir and readdir
For now, I'm just going off the `man` pages for these functions. Not
sure I'm in love with the ultra-terse style of K&R, with so much logic
taking place inside the condition checks of loops, but for now I'm going
for it as that's my only reference for "idiomatic".