Handle bad directory input
As much as possible, I'm trying to clone `ls`, so I took the same format
of error, which seems to be using just the standar library's `strerror`
function. Also returning 1 as that seems to be what `ls` is doing on my
machine, no matter the error, just playing around with it.