Use a struct to collect entries before printing
Because we have to potentially sort the list of entries, printing
one-by-one in the order given by `readdir` is not going to cut it.
Instead, allocate an array of custom structs that will be a small subset
of the `struct stat` returned by stat, only what's needed for printing.

For now I've arbitrarily set the number of files in the directory to
2^16 - 1, but it's probably worth revisting this decision. It would be
nice to actually implement as a stream but I'm not sure how to sort
that.