Change stat to fstatat to handle relative paths
The system call `stat` does not handle relative paths, so the program
would not work on anything other than the current directory. `fstatat`
handles this without having to build the full path yourself, as long as
you provide a file descriptor for the directory and some optional flags,
which I've set to 0.