Add a test for the buildSearchIndex function
This is a pretty complex test and I miss a lot of the conveniences other
languages have with dedicated frameworks. That said, I do appreciate
that Go forces you to think of your tests as just code, and code that's
a client of your program. Perhaps sometimes those shortcuts obfuscate
things that should be stated directly.